# Improve VSFS performance