--- a/include/Graphs/SVFG.h
+++ b/include/Graphs/SVFG.h
@@ -250,6 +250,9 @@ public:
         return nodeNum;
     }
 
+    /// Used *only* for Versioned FSPTA to encode propagation of versions
+    /// in the worklist (allowing for breadth-first propagation).
+    /// Returns the created node.
     inline const DummyVersionPropSVFGNode *addDummyVersionPropSVFGNode(const NodeID object, const NodeID version)
     {
         DummyVersionPropSVFGNode *dvpNode = new DummyVersionPropSVFGNode(totalVFGNode++, object, version);