This PR introduces two changes:

**BFS instead of DFS for address-taken propagation**
As discussed, propagation is now breadth-first. Since there is no way to add a _propagate o:v to o:v'_ constraint to the solver, I encode this as a dummy node, which holds this information and is pushed into the worklist as required. Otherwise, `propagateVersion` is no longer recursive and some `stmtReliance` stuff is pushed out to `processStore` (previously implied by `propagateVersion`).

**Ignoring existing constraints for call-graph updates**
The changes in `updateConnectedNode` ensure that we don't propagate `o:v` to `o:v'` if it has been done before. `o:v` is not necessarily new to require propagation if a reliance on `o:v` has previously been established.

`vis` is the smallest bitcode I noticed the regression in. It has (dense):
Old vfspta: 25s, new vfspta: 5s-6s, fspta: 15s
`bash` is the largest BC I have on hand that I could run on my 8 GB machine. It has (dense):
Old vfspta: 130s, new vfspta: 65s, fspta: starts swapping too much
A lot of smaller bitcode which I tested didn't really see any significant difference.

Unfortunately, with the server still down I can't test on larger bitcode.